+++ src/logging.c
@@ -121,7 +121,7 @@ void _log(const char *file, int line, int severity, const char *fmt, ...) {
   if (chars < 0 || chars >= LOG_LINE_SIZE/2) {
     abort();  // must be impossible
   }
-  buff_pos += chars;
+  buff_pos += (uint32_t)chars;
 
   va_list args;
   va_start(args, fmt);
@@ -132,7 +132,7 @@ void _log(const char *file, int line, int severity, const char *fmt, ...) {
   if (chars < 0) {
     abort();  // must be impossible
   }
-  buff_pos += chars;
+  buff_pos += (uint32_t)chars;
   if (buff_pos >= LOG_LINE_SIZE) {
     buff_pos = LOG_LINE_SIZE - 1;
     buff[buff_pos - 1] = '$'; // indicate truncation
